AppColors를 Colors로 이동하고 AppTypography를 이용한 AppTheme로 변경
